# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DriveLocal API
  version: 0.3.0
  description: |
    Backend for the single-driver ride-booking app serving the Kriel /
    Thubelihle area in ZAR. One owner-driver, booked by customers via SMS-OTP.

    Auth: request an OTP, then verify it to receive a `token` (JWT). Send the
    token as `Authorization: Bearer <token>` on role-gated routes. The owner
    driver's phone can also hit customer routes (to preview the app).
  contact:
    name: DriveLocal
servers:
  - url: http://localhost:4000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        code: { type: string, nullable: true }
...
    Point:
      type: object
      properties:
        lat: { type: number }
        lng: { type: number }
        address: { type: string, nullable: true }
        note: { type: string, nullable: true }
    Estimate:
      type: object
      properties:
        baseFare: { type: number }
        distanceCharge: { type: number }
        durationCharge: { type: number }
        subtotal: { type: number }
        distanceKm: { type: number }
        durationMin: { type: number }
        tipAmount: { type: number }
        discount: { type: number, description: "Promo discount applied (0 if none)" }
        promoCode: { type: string, nullable: true }
        promoPercent: { type: number, nullable: true }
        total: { type: number }
    EstimateBody:
      type: object
      required: [pickup, destination]
      properties:
        pickup: { $ref: '#/components/schemas/Point' }
        destination: { $ref: '#/components/schemas/Point' }
        promoCode: { type: string, nullable: true }
    Trip:
      type: object
      properties:
        id: { type: string }
        customerId: { type: string }
        driverId: { type: string, nullable: true }
        status:
          type: string
          enum: [requested, accepted, ongoing, completed, cancelled, scheduled]
        pickup: { $ref: '#/components/schemas/Point' }
        destination: { $ref: '#/components/schemas/Point' }
        priceModel: { type: string }
        paymentMethod: { type: string, enum: [cash, card] }
        distanceKm: { type: number }
        durationMin: { type: number }
        fareEstimate: { type: number }
        finalFare: { type: number, nullable: true }
        tipAmount: { type: number, default: 0 }
        rating: { type: integer, nullable: true }
        promoCode: { type: string, nullable: true }
        promoDiscount: { type: number, nullable: true }
        scheduledAt: { type: string, nullable: true, format: date-time }
        requestedAt: { type: string }
        timestamps: { type: object, additionalProperties: true }
    CreateTripBody:
      type: object
      required: [pickup, destination, paymentMethod]
      properties:
        pickup: { $ref: '#/components/schemas/Point' }
        destination: { $ref: '#/components/schemas/Point' }
        priceModel: { type: string, default: distance_time }
        paymentMethod: { type: string, enum: [cash, card] }
        scheduledAt: { type: string, nullable: true, description: "Future time → scheduled trip; omit for immediate" }
        promoCode: { type: string, nullable: true }
    Place:
      type: object
      properties:
        id: { type: string }
        label: { type: string }
        kind: { type: string, enum: [home, work, place] }
        address: { type: string, nullable: true }
        lat: { type: number }
        lng: { type: number }
        note: { type: string, nullable: true }
    RecentDestination:
      type: object
      properties:
        id: { type: string }
        destAddress: { type: string, nullable: true }
        destLat: { type: number }
        destLng: { type: number }
        destNote: { type: string, nullable: true }
        usedCount: { type: integer }
        lastUsed: { type: string }
    Dispute:
      type: object
      properties:
        id: { type: string }
        tripId: { type: string }
        userId: { type: string }
        reason: { type: string, nullable: true }
        status: { type: string, enum: [open, resolved] }
        resolution: { type: string, nullable: true }
        createdAt: { type: string }
        resolvedAt: { type: string, nullable: true }
...
paths:
...
  /api/customer/estimate:
    post:
      tags: [Customer]
      summary: Estimate fare (optionally with promo)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EstimateBody' }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  estimate: { $ref: '#/components/schemas/Estimate' }
                  route:
                    type: object
                    properties:
                      distanceKm: { type: number }
                      durationMin: { type: number }
                      polyline: { type: string, nullable: true }
  /api/customer/places:
    get:
      tags: [Customer]
      summary: List saved places (home first)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Place' }
    post:
      tags: [Customer]
      summary: Save a place
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Place' }
      responses:
        "201": { description: Created }
  /api/customer/places/{id}:
    put:
      tags: [Customer]
      summary: Update a saved place
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
    delete:
      tags: [Customer]
      summary: Delete a saved place
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
  /api/customer/trips:
    post:
      tags: [Customer]
      summary: Book a trip (immediate or scheduled)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTripBody' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  trip: { $ref: '#/components/schemas/Trip' }
                  estimate: { $ref: '#/components/schemas/Estimate' }
        "400": { description: Invalid payment / promo }
        "409":
          description: Driver offline or pickup out of service area
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      tags: [Customer]
      summary: Trip history
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Trip' }
  /api/customer/trips/active:
    get:
      tags: [Customer]
      summary: Current active trip (if any)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  trip: { $ref: '#/components/schemas/Trip', nullable: true }
  /api/customer/trips/upcoming:
    get:
      tags: [Customer]
      summary: Scheduled trips ahead
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  trips:
                    type: array
                    items: { $ref: '#/components/schemas/Trip' }
  /api/customer/trips/{id}/cancel:
    post:
      tags: [Customer]
      summary: Cancel a requested/accepted/scheduled trip
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, nullable: true }
      responses:
        "200": { description: OK }
  /api/customer/trips/{id}/rate:
    post:
      tags: [Customer]
      summary: Rate + optionally tip a completed trip
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stars: { type: integer, minimum: 1, maximum: 5 }
                tipAmount: { type: number, default: 0 }
      responses:
        "200": { description: OK }
  /api/customer/trips/{id}/fare-confirm:
    post:
      tags: [Customer]
      summary: Confirm the driver's final fare
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
  /api/customer/trips/{id}/dispute:
    post:
      tags: [Customer]
      summary: Open a fare dispute on a completed trip
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, nullable: true }
      responses:
        "201":
          description: Opened
          content:
            application/json:
              schema:
                type: object
                properties:
                  dispute: { $ref: '#/components/schemas/Dispute' }
        "409": { description: Trip not completed or dispute already open }
  /api/customer/recent-destinations:
    get:
      tags: [Customer]
      summary: Most-recent destinations for one-tap rebooking
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  destinations:
                    type: array
                    items: { $ref: '#/components/schemas/RecentDestination' }
  /api/customer/recent-destinations/{id}:
    delete:
      tags: [Customer]
      summary: Remove a recent destination
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
  /api/customer/push-token:
    post:
      tags: [Customer]
      summary: Register an FCM push token
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
                platform: { type: string }
      responses:
        "200": { description: OK }
...
  /api/driver/trips:
    get:
      tags: [Driver]
      summary: Driver trip history
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
...
  /api/driver/trips/{id}/cancel:
    post:
      tags: [Driver]
      summary: Cancel a trip
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
...